Refactor package location of aggregators
