+++ photon-api/src/main/scala/com/linkedin/photon/ml/SparkSessionConfiguration.scala
@@ -21,9 +21,9 @@ import org.apache.spark.serializer.KryoSerializer
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.SparkConf
 
+import com.linkedin.photon.ml.aggregators._
 import com.linkedin.photon.ml.data.{GameDatum, LabeledPoint, LocalDataset}
 import com.linkedin.photon.ml.function._
-import com.linkedin.photon.ml.function.glm.{HessianVectorAggregator, ValueAndGradientAggregator}
 import com.linkedin.photon.ml.model.Coefficients
 import com.linkedin.photon.ml.normalization.NormalizationContext
 import com.linkedin.photon.ml.optimization._
@@ -52,6 +52,8 @@ object SparkSessionConfiguration {
     classOf[GeneralizedLinearModel],
     classOf[GeneralizedLinearOptimizationProblem[_]],
     classOf[GLMOptimizationConfiguration],
+    classOf[HessianDiagonalAggregator],
+    classOf[HessianMatrixAggregator],
     classOf[HessianVectorAggregator],
     classOf[LinearSubspaceProjector],
     classOf[LabeledPoint],
